requirements.txt
requests>=2.22.0
dataclasses>=0.6
tox
pytest
python-dotenv
flit
boto3